big step - topic caching

+++ app/src/main/java/com/holdbetter/fintechchatproject/domain/entity/Topic.kt
@@ -6,7 +6,7 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 data class Topic(
     @field:Json(name = "max_id")
-    val maxId: Int,
+    val maxId: Long,
     @field:Json(name = "name")
     val name: String
 )
